wrapper: Install bin on first execution and remove hooks
Relying on installation hooks can cause installation issues (hooks can
be disabled!). Instead, the Changelog binary should be installed on
first execution.

First, we add a NodeJS executable script. This script attempts to
execute the Changelog binary, passing the arguments it received after
removing the args utilized by Node. If the Changelog binary is not
installed, this script attempts to download a precompiled binary,
falling back to compiling from local or repo source.

Next, we tell NPM to expect said script to be the `changelog` binary for
this package.

Lastly, we remove the existing hooks. We no longer need those now that
the binary known to NPM is a wrapper that can manage the installation of
the "real" Changelog binary.